// server/src/api/types.ts
export interface LoginUserRequest {
  email: string;
  password: string;
}

export interface LoginUserResponse {
  authToken: string;
  userId: string;
  name: string;
  role: Role;
}
// ...
export type Role = "client" | "realtor" | "admin";
// ...
export type LoginUser_Response = {
  kind: "success";
  data: LoginUserResponse;
} | {
  kind: "failure";
  data: string;
};